requirements.txt
flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
werkzeug==2.3.7
markdown==3.4.3
pydantic==2.1.0
